Refactor: 결합도 제거를 위한 Payment Enum 이관 및 Payload 변환 로직 적용
